fix export using cli

+++ lib/ad_localize/cli.rb
@@ -4,11 +4,14 @@ module AdLocalize
     def self.start(args:)
       options = OptionHandler.parse!(args)
       export_request = Mappers::OptionsToExportRequest.new.map(options: options)
+      LOGGER.debug("Export request options : #{export_request}")
       if export_request.has_sheets?
-        export_request.downloaded_csvs = DownloadSpreadsheets.new.call(export_request: export_request)
+        export_request.downloaded_csvs = Interactors::DownloadSpreadsheets.new.call(export_request: export_request)
       end
       Interactors::ProcessExportRequest.new.call(export_request: export_request)
     ensure
+      return unless export_request
+
       export_request.downloaded_csvs.each do |file|
         file.close
         file.unlink
